Update loop run
